Add minor refactoring
